// src/pages/detail-page/enum.ts
enum TagNames {
  DIV = 'div',
  NAV = 'nav',
  A = 'a',
  IMG = 'img',
  SPAN = 'span',
}

enum Styles {
  CONTAINER = 'detail-page',
  NAV = 'navbar-breadcrumb',
  PRODUCT_WRAPPER = 'product-container',
  SIDEBAR = 'sidebar-container',
  SLIDER = 'slider-container',
}

enum Events {
  CLICK = 'click',
}

enum Attributes {
  SRC = 'src',
  HREF = 'href',
}

export { TagNames, Styles, Events, Attributes };